# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * GatherSpace event venue marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type VenueAvailability = typeof VenueAvailability[keyof typeof VenueAvailability];


export const VenueAvailability = {
  open: 'open',
  limited: 'limited',
  booked: 'booked',
} as const;

export interface PolicyTier {
  label: string;
  refund: string;
  cutoff: string;
}

export interface CancellationPolicy {
  summary: string;
  tiers: PolicyTier[];
}

export interface Venue {
  id: string;
  name: string;
  area: string;
  city: string;
  capacity: number;
  startingPrice: number;
  rating: number;
  reviewCount: number;
  verified: boolean;
  imageUrl: string;
  amenities: string[];
  tags: string[];
  lat: number;
  lng: number;
  availability: VenueAvailability;
  description?: string;
  gallery?: string[];
  policy?: CancellationPolicy;
}

export type BiddingRequestStatus = typeof BiddingRequestStatus[keyof typeof BiddingRequestStatus];


export const BiddingRequestStatus = {
  open: 'open',
  reviewing: 'reviewing',
  awarded: 'awarded',
  expired: 'expired',
} as const;

export interface BiddingRequest {
  id: string;
  eventType: string;
  date: string;
  guests: number;
  area: string;
  budgetMin: number;
  budgetMax: number;
  amenities: string[];
  status: BiddingRequestStatus;
  bidCount: number;
  createdAt: string;
  note?: string;
}

export interface BiddingRequestInput {
  eventType: string;
  date: string;
  /** @minimum 1 */
  guests: number;
  /** @minLength 1 */
  area: string;
  /** @minimum 0 */
  budgetMin: number;
  /** @minimum 0 */
  budgetMax: number;
  amenities?: string[];
  note?: string;
}

export type BidStatus = typeof BidStatus[keyof typeof BidStatus];


export const BidStatus = {
  submitted: 'submitted',
  shortlisted: 'shortlisted',
  accepted: 'accepted',
  declined: 'declined',
} as const;

export interface Bid {
  id: string;
  requestId: string;
  venueId: string;
  venueName: string;
  amount: number;
  guestCount: number;
  inclusions: string[];
  status: BidStatus;
  createdAt: string;
  note?: string;
}

export type BiddingRequestDetail = BiddingRequest & {
  bids: Bid[];
};

export interface BidInput {
  venueId: string;
  /** @minimum 0 */
  amount: number;
  /** @minimum 1 */
  guestCount: number;
  inclusions: string[];
  note?: string;
}

export type AvailabilityResultStatus = typeof AvailabilityResultStatus[keyof typeof AvailabilityResultStatus];


export const AvailabilityResultStatus = {
  available: 'available',
  reserved: 'reserved',
  confirmed: 'confirmed',
} as const;

export interface AvailabilityResult {
  venueId: string;
  date: string;
  slot: string;
  status: AvailabilityResultStatus;
  message?: string;
}

export interface BookingInput {
  venueId: string;
  date: string;
  slot: string;
  /** @minimum 1 */
  guests: number;
  /** @minimum 0 */
  amount: number;
  liabilityAccepted: boolean;
  bidId?: string;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  pending_payment: 'pending_payment',
  payment_review: 'payment_review',
  confirmed: 'confirmed',
  cancelled: 'cancelled',
} as const;

export interface Booking {
  id: string;
  venueId: string;
  date: string;
  slot: string;
  guests: number;
  amount: number;
  status: BookingStatus;
  liabilityAccepted: boolean;
  createdAt: string;
}

export type MessageSender = typeof MessageSender[keyof typeof MessageSender];


export const MessageSender = {
  client: 'client',
  ai: 'ai',
  manager: 'manager',
} as const;

export interface Message {
  id: string;
  conversationId: string;
  sender: MessageSender;
  text: string;
  createdAt: string;
}

export interface MessageInput {
  /** @minLength 1 */
  text: string;
}

export type HandoffStatus = typeof HandoffStatus[keyof typeof HandoffStatus];


export const HandoffStatus = {
  ai: 'ai',
  waiting_for_manager: 'waiting_for_manager',
  manager: 'manager',
} as const;

export interface Handoff {
  conversationId: string;
  status: HandoffStatus;
  message: string;
}

export interface OwnerSummary {
  venueName: string;
  profileComplete: number;
  monthlyViews: number;
  openRequests: number;
  upcomingBookings: number;
  pendingRevenue: number;
  recentActivity: string[];
}

export type EventTypeParameter = typeof EventTypeParameter[keyof typeof EventTypeParameter];


export const EventTypeParameter = {
  wedding: 'wedding',
  birthday: 'birthday',
  corporate: 'corporate',
  shower: 'shower',
  other: 'other',
} as const;

export type AreaParameter = string;

export type GuestCountParameter = number;

export type BudgetParameter = number;

export type ListVenuesParams = {
eventType?: EventTypeParameter;
area?: AreaParameter;
/**
 * @minimum 1
 */
guestCount?: GuestCountParameter;
/**
 * @minimum 0
 */
budget?: BudgetParameter;
};

export type CheckAvailabilityParams = {
venueId: string;
date: string;
slot: CheckAvailabilitySlot;
};

export type CheckAvailabilitySlot = typeof CheckAvailabilitySlot[keyof typeof CheckAvailabilitySlot];


export const CheckAvailabilitySlot = {
  morning: 'morning',
  evening: 'evening',
  full_day: 'full_day',
} as const;