# lib/api-zod/src/generated/types/biddingRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * GatherSpace event venue marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { BiddingRequestStatus } from './biddingRequestStatus';

export interface BiddingRequest {
  id: string;
  eventType: string;
  date: Date;
  guests: number;
  area: string;
  budgetMin: number;
  budgetMax: number;
  amenities: string[];
  status: BiddingRequestStatus;
  bidCount: number;
  createdAt: Date;
  note?: string;
}