# lib/api-zod/src/generated/types/biddingRequestDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * GatherSpace event venue marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Bid } from './bid';
import type { BiddingRequest } from './biddingRequest';

export type BiddingRequestDetail = BiddingRequest & {
  bids: Bid[];
};