# lib/api-zod/src/generated/types/biddingRequestInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * GatherSpace event venue marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface BiddingRequestInput {
  eventType: string;
  date: Date;
  /** @minimum 1 */
  guests: number;
  /** @minLength 1 */
  area: string;
  /** @minimum 0 */
  budgetMin: number;
  /** @minimum 0 */
  budgetMax: number;
  amenities?: string[];
  note?: string;
}