# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * GatherSpace event venue marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: string;
  venueId: string;
  date: Date;
  slot: string;
  guests: number;
  amount: number;
  status: BookingStatus;
  liabilityAccepted: boolean;
  createdAt: Date;
}