# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * GatherSpace event venue marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  venueId: string;
  date: Date;
  slot: string;
  /** @minimum 1 */
  guests: number;
  /** @minimum 0 */
  amount: number;
  liabilityAccepted: boolean;
  bidId?: string;
}